requirements.txt
anthropic==0.34.1
python-dotenv==1.0.1
openai==1.35.13
Pillow